QR Item generation
When you buy a product, it generates a QR Item that is displayed in "My purchases" if it's not used and in "My History" if it is.
TO DO: Display view of item when you scan the QR with a buttom to use it.
TO DO: Generate the url in the QR displayed for every purchase.